use user_id in orders/payments

+++ db/orders_migrate/20201011085636_create_order_lines.rb
@@ -2,7 +2,7 @@ class CreateOrderLines < ActiveRecord::Migration[7.0]
   def change
     create_table :order_lines do |t|
       t.references :order, null: false, foreign_key: true
-      t.integer :product_id, null: false
+      t.bigint :product_id, null: false
       t.integer :quantity, null: false
       t.boolean :reserved, null: false, default: false
       t.integer :price_at_submit
